fix: bugfix on nextauth route exports
